ISSUE #4587 - improve baseCalendarPicker and replace baseDueDate with it
