+++ frontend/src/v5/ui/routes/viewer/tickets/ticketsList/ticketItem/ticketItem.component.tsx
@@ -54,7 +54,7 @@ export const TicketItem = ({ ticket, onClick, selected }: TicketItemProps) => {
 	const onBlurAssignees = (newVals) => {
 		if (!isEqual(newVals, assignees)) updateTicketProperty({ [IssueProperties.ASSIGNEES]: newVals });
 	};
-	const onBlurDueDate = (newVal) => {
+	const onChangeDueDate = (newVal) => {
 		if (newVal !== dueDate) updateTicketProperty({ [IssueProperties.DUE_DATE]: newVal });
 	};
 
@@ -85,7 +85,7 @@ export const TicketItem = ({ ticket, onClick, selected }: TicketItemProps) => {
 			</ChipList>
 			{priority && (
 				<IssuePropertiesRow>
-					<DueDateWithLabel value={dueDate} onBlur={onBlurDueDate} disabled={readOnly} />
+					<DueDateWithLabel value={dueDate} onChange={onChangeDueDate} disabled={readOnly} />
 					<Chip {...PRIORITY_LEVELS_MAP[priority]} variant="text" label="" />
 					<Assignees value={assignees} onBlur={onBlurAssignees} disabled={readOnly} />
 				</IssuePropertiesRow>
